+++ R/lccWrapper.R
@@ -38,5 +38,7 @@ lccWrapper <- function(model, q_f, tk, diffbeta, n.delta) {
                      delta = delta, deltal = deltal, model = model)
   if(length(rho)==1){
     return(rho[[1]])
-  }else(return(rho[[n.delta]]))
+  }else(if(is.na(rho[[2]])){
+    return(rho[[1]])
+  }else(return(rho[[n.delta]])))
 }