+++ R/plotControl.R
@@ -15,33 +15,59 @@
 #######################################################################
 ##' @title Specifying graphical control values for \code{lcc} class
 ##'
-##' @description The values supplied in the \code{plotControl()} call replace the defaults, and a \code{\link{list}} with all settings is returned.
+##' @description The values supplied in the \code{plotControl()} call
+##'   replace the defaults, and a \code{\link{list}} with all settings
+##'   is returned.
 ##'
 ##' @return a list with components for each of the possible arguments.
 ##'
-##' @param plot an optional to include an initial plot. If \code{TRUE}, the default, returns a \code{\link[ggplot2]{ggplot}} object with a initial plot for \code{lcc} class. If \code{FALSE} never includes.
+##' @param plot an optional to include an initial plot. If \code{TRUE},
+##'   the default, returns a \code{\link[ggplot2]{ggplot}} object with a
+##'   initial plot for \code{lcc} class. If \code{FALSE} never includes.
 ##'
-##' @param shape Draw points considering a shape parameter. Legal shape values are the numbers 0 to 25, and 32 to 127; see \code{\link[ggplot2]{aes_linetype_size_shape}}. Default is \code{1}.
+##' @param shape Draw points considering a shape parameter. Legal shape
+##'   values are the numbers 0 to 25, and 32 to 127; see
+##'   \code{\link[ggplot2]{aes_linetype_size_shape}}. Default is
+##'   \code{1}.
 ##'
-##' @param colour an specification for lines color. Default is \code{"black"}.
+##' @param colour an specification for lines color. Default is
+##'   \code{"black"}.
 ##'
-##' @param size an specification for lines size. Should be specified with a numerical value (in millimetres); see \code{\link[ggplot2]{aes_linetype_size_shape}}. Default is \code{0.5}.
+##' @param size an specification for lines size. Should be specified
+##'   with a numerical value (in millimetres); see
+##'   \code{\link[ggplot2]{aes_linetype_size_shape}}. Default is
+##'   \code{0.5}.
 ##'
-##' @param xlab a title for the \code{x} axis.  Default is \code{"Time"}.
+##' @param xlab a title for the \code{x} axis.  Default is
+##'   \code{"Time"}.
 ##'
-##' @param LCC_ylab title for the \code{y} axis related with longitudinal concordance correlation (LCC). Default is \code{"LCC"}.
+##' @param LCC_ylab title for the \code{y} axis related with
+##'   longitudinal concordance correlation (LCC). Default is
+##'   \code{"LCC"}.
 ##'
-##' @param LPC_ylab title for the \code{y} axis related with longitudinal Pearson correlation (LPC). Default is \code{"LPC"}.
+##' @param LPC_ylab title for the \code{y} axis related with
+##'   longitudinal Pearson correlation (LPC). Default is \code{"LPC"}.
 ##'
-##' @param LA_ylab title for the \code{y} axis related with longitudinal accuracy (LA). Default is \code{"LA"}.
+##' @param LA_ylab title for the \code{y} axis related with longitudinal
+##'   accuracy (LA). Default is \code{"LA"}.
 ##'
-##' @param LCC_scale_y_continuous a numeric vector of length two providing limits of the scale related with longitudinal concordance correlation (LCC). Default is \code{c(0,1)}.
+##' @param LCC_scale_y_continuous a numeric vector of length two
+##'   providing limits of the scale related with longitudinal
+##'   concordance correlation (LCC). Default is \code{c(0,1)}.
 ##'
-##' @param LPC_scale_y_continuous a numeric vector of length two providing limits of the scale related with longitudinal concordance correlation (LCC). Default is \code{c(0,1)}.
+##' @param LPC_scale_y_continuous a numeric vector of length two
+##'   providing limits of the scale related with longitudinal
+##'   concordance correlation (LCC). Default is \code{c(0,1)}.
 ##'
-##' @param LA_scale_y_continuous a numeric vector of length two providing limits of the scale related with longitudinal concordance correlation (LCC). Default is \code{c(0,1)}.
+##' @param LA_scale_y_continuous a numeric vector of length two
+##'   providing limits of the scale related with longitudinal
+##'   concordance correlation (LCC). Default is \code{c(0,1)}.
 ##'
-##' @param all.plot viewports functions for \code{lcc} class. If \code{TRUE}, the default, returns a object created by \code{\link[grid]{viewport}} function with multiple plots in a single page. If \code{FALSE} returns a single \code{\link[ggplot2]{ggplot}} object by page.
+##' @param all.plot viewports functions for \code{lcc} class. If
+##'   \code{TRUE}, the default, returns a object created by
+##'   \code{\link[grid]{viewport}} function with multiple plots in a
+##'   single page. If \code{FALSE} returns a single
+##'   \code{\link[ggplot2]{ggplot}} object by page.
 ##'
 ##' @author Thiago de Paula Oliveira, \email{thiago.paula.oliveira@@usp.br}
 ##'