+++ R/plotmethod_lcc.R
@@ -14,48 +14,81 @@
 #                                                                     #
 #######################################################################
 ##' @title Plot an lcc object
-##' 
-##' @description A plot of predictions versus the time covariate is generated. Predicted values are joined by lines while sampled observations are represented by circles. If the argument \code{components=TRUE} is considered in the \code{lcc} object, single plots of each statistics are returned on differents pages.           
+##'
+##' @description A plot of predictions versus the time covariate is
+##'   generated. Predicted values are joined by lines while sampled
+##'   observations are represented by circles. If the argument
+##'   \code{components=TRUE} is considered in the \code{lcc} object,
+##'   single plots of each statistics are returned on differents pages.
 ##'
 ##' @usage
 ##' lccPlot(obj, control, ...)
-##'     
-##' @param obj an object inheriting from class "lcc", representing a fitted lcc model.
-##' 
-##' @param control a list of control values or character strings returned by the function \code{\link{plotControl}}. Defaults to an empty list.  The list may contain the following components:
-##' \describe{
-##' \item{\code{shape}:}{draw points considering a shape parameter. Possible shape values are the numbers 0 to 25, and 32 to 127; see \code{\link[ggplot2]{aes_linetype_size_shape}}. Default is \code{1}.}
 ##'
-##' \item{\code{colour}:}{specification for lines color. Default is \code{"black"}.}
+##' @param obj an object inheriting from class "lcc", representing a
+##'   fitted lcc model.
 ##'
-##' \item{\code{size}:}{specification for lines size. Should be specified with a numerical value (in millimetres); see \code{\link[ggplot2]{aes_linetype_size_shape}}. Default is \code{0.5}.}
+##' @param control a list of control values or character strings
+##'   returned by the function \code{\link{plotControl}}. Defaults to an
+##'   empty list.  The list may contain the following components:
+##'   \describe{ \item{\code{shape}:}{draw points considering a shape
+##'   parameter. Possible shape values are the numbers 0 to 25, and 32
+##'   to 127; see
+##'   \code{\link[ggplot2]{aes_linetype_size_shape}}. Default is
+##'   \code{1}.}
 ##'
-##' \item{\code{xlab}:}{title for the \code{x} axis.  Default is \code{"Time"}.}
+##' \item{\code{colour}:}{specification for lines color. Default is
+##' \code{"black"}.}
 ##'
-##' \item{\code{LCC_ylab}:}{title for the \code{y} axis related with LCC. Default is \code{"LCC"}.}
+##' \item{\code{size}:}{specification for lines size. Should be
+##' specified with a numerical value (in millimetres); see
+##' \code{\link[ggplot2]{aes_linetype_size_shape}}. Default is
+##' \code{0.5}.}
 ##'
-##' \item{\code{LPC_ylab}:}{title for the \code{y} axis related with LPC. Default is \code{"LPC"}.}
+##' \item{\code{xlab}:}{title for the \code{x} axis.  Default is
+##' \code{"Time"}.}
 ##'
-##' \item{\code{LA_ylab}:}{title for the \code{y} axis related with LA. Default is \code{"LA"}.}
+##' \item{\code{LCC_ylab}:}{title for the \code{y} axis related with
+##' LCC. Default is \code{"LCC"}.}
 ##'
-##' \item{\code{LCC_scale_y_continuous}:}{numeric vector of length two providing limits of the scale related to LCC. Default is \code{c(0,1)}.}
+##' \item{\code{LPC_ylab}:}{title for the \code{y} axis related with
+##' LPC. Default is \code{"LPC"}.}
 ##'
-##' \item{\code{LPC_scale_y_continuous}:}{numeric vector of length two providing limits of the scale related to LPC. Default is \code{c(0,1)}.}
+##' \item{\code{LA_ylab}:}{title for the \code{y} axis related with
+##' LA. Default is \code{"LA"}.}
 ##'
-##' \item{\code{LA_scale_y_continuous}:}{numeric vector of length two providing limits of the scale related to LA. Default is \code{c(0,1)}.}
+##' \item{\code{LCC_scale_y_continuous}:}{numeric vector of length two
+##' providing limits of the scale related to LCC. Default is
+##' \code{c(0,1)}.}
 ##'
-##' \item{\code{all.plot}:}{\code{viewport} functions for the \code{lcc} class. If \code{TRUE}, the default, returns an object created by the \code{\link[grid]{viewport}} function with multiple plots on a single page. If \code{FALSE} returns a single \code{\link[ggplot2]{ggplot}} object by different pages.}
-##' }
-##' 
+##' \item{\code{LPC_scale_y_continuous}:}{numeric vector of length two
+##' providing limits of the scale related to LPC. Default is
+##' \code{c(0,1)}.}
+##'
+##' \item{\code{LA_scale_y_continuous}:}{numeric vector of length two
+##' providing limits of the scale related to LA. Default is
+##' \code{c(0,1)}.}
+##'
+##' \item{\code{all.plot}:}{\code{viewport} functions for the \code{lcc}
+##' class. If \code{TRUE}, the default, returns an object created by the
+##' \code{\link[grid]{viewport}} function with multiple plots on a
+##' single page. If \code{FALSE} returns a single
+##' \code{\link[ggplot2]{ggplot}} object by different pages.}  }
+##'
 ##' @param ... not used.
-##' 
+##'
 ##' @author Thiago de Paula Oliveira, \email{thiago.paula.oliveira@@usp.br}
-##' 
-##' @references Lin, L. A Concordance Correlation Coefficient to Evaluate Reproducibility. \emph{Biometrics}, 45, n. 1, 255-268, 1989. 
-##' @references Oliveira, T.P.; Hinde, J.; Zocchi S.S. Longitudinal Concordance Correlation Function Based on Variance Components: An Application in Fruit Color Analysis. \emph{Journal of Agricultural, Biological, and Environmental Statistics}, v. 23, n. 2, 233–254, 2018.
-##' 
+##'
+##' @references Lin, L. A Concordance Correlation Coefficient to
+##'   Evaluate Reproducibility. \emph{Biometrics}, 45, n. 1, 255-268,
+##'   1989.
+##' @references Oliveira, T.P.; Hinde, J.; Zocchi S.S. Longitudinal
+##'   Concordance Correlation Function Based on Variance Components: An
+##'   Application in Fruit Color Analysis. \emph{Journal of
+##'   Agricultural, Biological, and Environmental Statistics}, v. 23,
+##'   n. 2, 233–254, 2018.
+##'
 ##' @seealso \code{\link[lcc]{lcc}}.
-##' 
+##'
 ##' @examples
 ##'
 ##' data(hue)
@@ -65,11 +98,11 @@
 ##'          method = "Method", time = "Time", qf = 2, qr = 2,
 ##'          components=TRUE)
 ##' lccPlot(fm1)
-##'          
+##'
 ##' @export
 
 lccPlot<-function(obj, control = list(), ...){
-  if(class(obj)!="lcc") stop("Object must inherit from class \"lcc\"", call.=FALSE) 
+  if(class(obj)!="lcc") stop("Object must inherit from class \"lcc\"", call.=FALSE)
   # Arguments for the plot
   plot.cons<-plotControl(shape=1, colour="black",
     size=0.5, xlab = "Time",
@@ -95,7 +128,7 @@ lccPlot<-function(obj, control = list(), ...){
       plot.cons[[pos[i]]]<-control[[i]]
     }
   }
-  
+
 
   #Standard arguments
   nd<-obj$plot_info$nd
@@ -105,7 +138,7 @@ lccPlot<-function(obj, control = list(), ...){
   ldb<-obj$plot_info$ldb
   ci<-obj$plot_info$ci
   components<-obj$plot_info$components
-  
+
     if(ci==FALSE) {
       if(ldb == 1) {
           plot_lcc(rho=obj$plot_info$rho, tk.plot= tk.plot,