+++ NEWS.md
@@ -2,13 +2,17 @@
 
 ## New argument
 
-*  We included an argument called "interaction " in the lcc function as an option to estimate the interaction effect between time and method. If interaction is declared FALSE, the lcc function consider only the main effects of time and method in the model.
+*  We included an argument called "interaction " in the lcc function as
+   an option to estimate the interaction effect between time and
+   method. If interaction is declared FALSE, the lcc function consider
+   only the main effects of time and method in the model.
 
 ## Bug fixes and minor improvements
 
-* Fixed issue when the wrong name of variable is declared.  
+* Fixed issue when the wrong name of variable is declared.
 
-* Fixed issue when three or more levels of methods is considered in the dataset.
+* Fixed issue when three or more levels of methods is considered in the
+  dataset.
 
 * Fixed issue in lccWrapper function get the correct rho:
   	* previous code: